Move logging tests to separate file
